+++ spec/lightstreamer/stream_connection_spec.rb
@@ -4,27 +4,16 @@ describe Lightstreamer::StreamConnection do
                                               password: 'password', adapter_set: 'set'
   end
 
-  let(:create_request) { instance_double 'Typhoeus::Request' }
-  let(:create_response) { instance_double 'Typhoeus::Response' }
-  let(:create_params) do
-    { LS_op2: 'create', LS_cid: 'mgQkwtwdysogQz2BJ4Ji kOj2Bg', LS_user: 'username', LS_password: 'password',
-      LS_adapter_set: 'set' }
-  end
+  let(:create_args) do
+    query = { LS_op2: 'create', LS_cid: 'mgQkwtwdysogQz2BJ4Ji kOj2Bg', LS_user: 'username', LS_password: 'password',
+              LS_adapter_set: 'set' }
 
-  before do
-    expect(Typhoeus::Request).to receive(:new)
-      .with('http://test.com/lightstreamer/create_session.txt', method: :post, params: create_params,
-                                                                connecttimeout: 15)
-      .and_return(create_request)
+    ['http://test.com/lightstreamer/create_session.txt', hash_including(query: query, connect_timeout: 15)]
   end
 
   it 'creates and runs a stream connection with a simple lifecycle' do
-    on_body_block = nil
-
-    expect(create_request).to receive(:on_body) { |&block| on_body_block = block }
-    expect(create_request).to receive(:on_complete)
-    expect(create_request).to receive(:run) do
-      on_body_block.call "OK\r\nSessionId:A\r\n\r\none\r\ntwo\r\n"
+    expect(Excon).to receive(:post).with(*create_args) do |_url, params|
+      params[:response_block].call "OK\r\nSessionId:A\r\n\r\none\r\ntwo\r\n", nil, nil
       loop {}
     end
 
@@ -42,29 +31,17 @@ describe Lightstreamer::StreamConnection do
   end
 
   it 'creates and runs a stream connection which rebinds itself in response to a LOOP message' do
-    bind_request = instance_double 'Typhoeus::Request'
-    bind_params = { LS_session: 'A' }
-
-    create_request_on_body_block = nil
-
-    expect(create_request).to receive(:on_body) { |&block| create_request_on_body_block = block }
-    expect(create_request).to receive(:on_complete)
-    expect(create_request).to receive(:run) do
-      create_request_on_body_block.call "OK\r\nSessionId:A\r\n\r\none\r\ntwo\r\nLOOP\r\n"
-    end
-
-    expect(Typhoeus::Request).to receive(:new)
-      .with('http://test.com/lightstreamer/bind_session.txt', method: :post, params: bind_params, connecttimeout: 15)
-      .and_return(bind_request)
-
-    bind_request_on_body_block = nil
     stream_thread = nil
 
-    expect(bind_request).to receive(:on_body) { |&block| bind_request_on_body_block = block }
-    expect(bind_request).to receive(:on_complete)
-    expect(bind_request).to receive(:run) do
+    expect(Excon).to receive(:post).with(*create_args) do |_url, params|
       stream_thread = Thread.current
-      bind_request_on_body_block.call "OK\r\nSessionId:A\r\n\r\nthree\r\nfour\r\n"
+      params[:response_block].call "OK\r\nSessionId:A\r\n\r\none\r\ntwo\r\nLOOP\r\n", nil, nil
+    end
+
+    bind_args = ['http://test.com/lightstreamer/bind_session.txt',
+                 hash_including(query: { LS_session: 'A' }, connect_timeout: 15)]
+    expect(Excon).to receive(:post).with(*bind_args) do |_url, params|
+      params[:response_block].call "OK\r\nSessionId:A\r\n\r\nthree\r\nfour\r\n", nil, nil
       sleep
     end
 
@@ -84,16 +61,13 @@ describe Lightstreamer::StreamConnection do
   end
 
   it 'creates and runs a stream connection which is terminated by the server' do
-    on_body_block = nil
     stream_thread = nil
 
-    expect(create_request).to receive(:on_body) { |&block| on_body_block = block }
-    expect(create_request).to receive(:on_complete)
-    expect(create_request).to receive(:run) do
+    expect(Excon).to receive(:post).with(*create_args) do |_url, params|
       stream_thread = Thread.current
-      on_body_block.call "OK\r\nSessionId:A\r\n\r\n"
+      params[:response_block].call "OK\r\nSessionId:A\r\n\r\n", nil, nil
       sleep
-      on_body_block.call "one\r\ntwo\r\nEND\r\nthree\r\n"
+      params[:response_block].call "one\r\ntwo\r\nEND\r\nthree\r\n", nil, nil
     end
 
     stream_connection = Lightstreamer::StreamConnection.new session
@@ -110,20 +84,11 @@ describe Lightstreamer::StreamConnection do
   end
 
   it 'handles an HTTP error on the stream thread' do
-    on_complete_block = nil
-
-    expect(create_request).to receive(:on_body)
-    expect(create_request).to receive(:on_complete) { |&block| on_complete_block = block }
-    expect(create_request).to receive(:run) { on_complete_block.call create_response }
-
-    expect(create_response).to receive(:success?).and_return(false)
-    expect(create_response).to receive(:return_message).and_return('Error message')
-    expect(create_response).to receive(:response_code).and_return(404)
+    expect(Excon).to receive(:post).and_raise(Excon::Error.new('message'))
 
     expect { Lightstreamer::StreamConnection.new(session).connect }.to raise_error do |error|
-      expect(error).to be_a(Lightstreamer::Errors::RequestError)
-      expect(error.request_error_message).to eq('Error message')
-      expect(error.request_error_code).to eq(404)
+      expect(error).to be_a(Lightstreamer::Errors::ConnectionError)
+      expect(error.message).to eq('message')
     end
   end
 end